user-service/requirements.txt
flask
prometheus-flask-exporter
psycopg2-binary
redis
python-dotenv